Move DAO and tables to agent-core
