+++ agent-core/build.yaml
@@ -1,0 +1,7 @@
+targets:
+  $default:
+    builders:
+      drift_dev:
+        options:
+          store_date_time_values_as_text: false
+          use_data_class_name_for_companions: true
