+++ agent-core/lib/src/database/converters/tool_call_converter.dart
