+++ agent-core/lib/src/database/daos/project_dao.dart
