+++ agent-core/lib/src/database/daos/session_dao.dart
