+++ agent-core/lib/src/database/tables/messages_table.dart
