+++ agent-core/lib/src/database/tables/projects_table.dart
