+++ agent-core/lib/src/database/tables/sessions_table.dart
