+++ agent-core/lib/src/database/tables/snippets_table.dart
