+++ decamp-app/lib/components/rich_message_content.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:gpt_markdown/gpt_markdown.dart';
 import 'package:intl/intl.dart';
-import 'package:decamp/database/database.dart';
+import 'package:agent_core/agent_core.dart';
 import 'package:decamp/utils/message_formatter.dart';
 import 'package:decamp/themes/terminal_theme.dart';
 import 'package:decamp/components/expandable_code_block.dart';
