+++ decamp-app/lib/pages/main_settings.dart
@@ -9,7 +9,6 @@ import '../providers/project_provider.dart';
 import '../providers/llm_settings_provider.dart';
 import '../providers/ssh_settings_provider.dart';
 import '../providers/settings_provider.dart';
-import '../database/database.dart';
 import '../utils/project_utils.dart';
 import '../components/ai_model_dialog.dart';
 import '../components/ssh_settings_dialog.dart';
