+++ decamp-app/lib/providers/message_provider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../database/database.dart';
+import 'package:agent_core/agent_core.dart';
 import 'database_provider.dart';
 import 'session_provider.dart';
