+++ decamp-app/lib/providers/session_provider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../database/database.dart';
+import 'package:agent_core/agent_core.dart';
 import 'database_provider.dart';
 import 'project_provider.dart';
