+++ decamp-app/lib/providers/snippet_provider.dart
@@ -1,7 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:drift/drift.dart';
 import 'package:agent_core/agent_core.dart';
-import '../database/database.dart';
 import '../providers/database_provider.dart';
 
 /// Stream provider for global snippets
